Added MainMenuState and functionality.
The MainMenuState prints out the main menu and asks the user for an input.
Currently, no other states exists so other states cannot be pushed onto the stack only removed so for now only the exit option works to remove the state.
All other options do nothing and the program repeats itself.