Created Accounts State which will lead to create account and delete account.
The main menu state now also allows option 3 to be selected and the account state is pushed onto the stack.
The Divider between states has been moved to the execute function as initialize was only called once on creation not every time the state is run.
Also one new line character has been removed from the divider at the start as after an input from the user a newline is added to the console anyway.