Created 4 new placeholder states.
Created state classes for: CreateAccount, DeleteAccount, Leaderboard, Play.
These placeholders only have one option which is to go back with option 1 and that pops the state from the stack.
These placeholders will be filled with functionality later but for now only are there to make sure that the navigation system works.