Created a serialize function and a constructor which takes in binary data.
The serialize function quantizes the entire Neural Network into a char array which will be used to save the NN to a file.
The new constructor does the inverse, it takes a char array of quantized data and reconstructs the Neural Network to the state it was in before serialization.

This has been tested to work as a new neural network can be created and produce the same output when given the same input. This was a great feeling to see the same number printed twice.