Added functionality to the save to file method.
The save to file saves all data about the account object into a file.
This includes the neural network data which uses the serialize method.
One problem is that ofstream will not create a directory, meaning the res/Accounts/Users/ directory must be created manually until I find a work around.
This also cannot be tested properly until the load from file method is created which will test that all data is stored and read in correctly.