Added functionality to load from file constructor.
This adds the ability to construct an Account object from a file with all data reminaing intact between saving and restoring.
This verifies that saving to the file works as all data matches before and after saving and loading to files.

One problem is still the directory not being created by fstream. As this is a prototype there is also no error checking making sure the file opened correctly